Fixed small port connection issue.
Fixed the following failed assert: "AutoRouterPort.createStartEndPointTo: isRightAngle( this.port_OnWhichEdge(resultPoint) FAILED". This was caused by full ports pushing an edge to the corner of a port; the path cannot start on the corner of a port because it is unclear what side of the port it is on.


Former-commit-id: 46ee215a824f389b3b21dc629383811ac0207a6f