Fixed initialization to occur on creation not first render
The `sourceCoordinates` value needs to be initialized before the
DiagramDesigner tries to finish it's drawing cycle. As the Autorouter
web worker is async, this means that the connection may not be
initialized at the first drawing cycle following it's creation. This
results in errors.

It now initializes the given values on creation rather than simply on
first drawing.


Former-commit-id: f97ab37556a59f7a876e961c978771a319e1acd5