Added optional requirement of responses in web worker. Fixes #333
Added respondToAll to the web worker so we could get feedback in the
testing tools and force it to behave synchronously. That is, in the
testing tools for the autorouter, it waits for a response from the web
worker before sending the next command; although it is inefficient, it
was important for getting feeback about certain things (such as error
checking as the web worker will send back an error if one occurs).


Former-commit-id: 6aa47cf5cc71b23f837d609dc16c51f81bf04f69