Option to disabled the use of user-defined/dynamic SVGs (#1448)
By default the dynamic svg usage is allowed, but as it poses security vulnerability it could be disabled.
When disabled, the user is not allowed to use the SVG text editor and if the svg container registry has some textual content, it will not be run/compiled/rendered. Only relative pathes can be used in this mode.

Former-commit-id: 586e5aef9be3f54ca3e6cd52ba73558c81729f23